requirements.txt
psycopg2
django==1.7.3
Pillow==2.7.0
django-photologue==3.1.1
django-pagedown
django-widget-tweaks
django-markdown-newsletter
markdown
